My Room List function finish

+++ src/controllers/roomController.js
@@ -1,12 +1,6 @@
 import Room from "../models/Room";
 import User from "../models/User";
 /*
-Room.find({}, (error, rooms) => {
-  if(error) {
-    return res.render("serveer-error");
-  } 
-  return res.render("home", { pageTitle: "Home", rooms });
-});
 callback: 에러를 바로 볼 수 있지만 기다리는 기능이 없어서 DB에서 불러오는
 시간이 있어서 순서가 꼬인다
 await, async : DB에게 결과 값을 받을 때까지 js가 기다리고 매우 직관적이다.
@@ -16,6 +10,9 @@ async function 선언은 AsyncFunction객체를 반환하는 하나의 비동기
 사용하여 결과를 반환합니다. 그러나 비동기 함수를 사용하는 코드의 구문과 구조는,
 표준 동기 함수를 사용하는것과 많이 비슷합니다.
 */
+const HTTP_FORBIDDEN = 403;
+const HTTP_PAGE_NOT_FOUND = 404;
+
 // globalRouter
 export const home = async (req, res) => {
   const rooms = await Room.find({}).sort({ createdAt: "desc" });
@@ -47,16 +44,26 @@ export const detail = async (req, res) => {
   const { id } = req.params;
   const room = await Room.findById(id).populate("host");
   if (!room) {
-    return res.status(404).render("404", { pageTitle: "Room not found." });
+    return res
+      .status(HTTP_PAGE_NOT_FOUND)
+      .render("404", { pageTitle: "Room not found." });
   }
   return res.render("rooms/detail", { pageTitle: room.title, room });
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const room = await Room.findById(id);
   if (!room) {
-    return res.status(404).render("404", { pageTitle: "Room not found." });
+    return res
+      .status(HTTP_PAGE_NOT_FOUND)
+      .render("404", { pageTitle: "Room not found." });
+  }
+  if (String(room.host) !== req.session.user._id) {
+    return res.status(HTTP_FORBIDDEN).redirect("/");
   }
   return res.render("rooms/editRoom", {
     pageTitle: `Edit : ${room.title}`,
@@ -65,11 +72,20 @@ export const getEdit = async (req, res) => {
 };
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, city, price, amenities } = req.body;
-  const room = await Room.exists({ _id: id });
+  const room = await Room.findById(id);
   if (!room) {
-    return res.status(404).render("404", { pageTitle: "Room not found." });
+    return res
+      .status(HTTP_PAGE_NOT_FOUND)
+      .render("404", { pageTitle: "Room not found." });
+  }
+  console.log(String(room.host));
+  if (String(room.host) !== req.session.user._id) {
+    return res.status(HTTP_FORBIDDEN).redirect("/");
   }
   await Room.findByIdAndUpdate(id, {
     title,
@@ -93,7 +109,7 @@ export const postUpload = async (req, res) => {
   const { path: fileUrl } = req.file;
   const { title, description, city, price, amenities } = req.body;
   try {
-    await Room.create({
+    const newRoom = await Room.create({
       title,
       description,
       city,
@@ -102,6 +118,9 @@ export const postUpload = async (req, res) => {
       fileUrl,
       host: _id,
     });
+    const user = await User.findById(_id);
+    user.rooms.push(newRoom._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     console.log("error : ", error);
@@ -114,7 +133,22 @@ export const postUpload = async (req, res) => {
 
 export const deleteRoom = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const room = await Room.findById(id);
+  const user = await Room.findById(_id);
+  if (!room) {
+    return res
+      .status(HTTP_PAGE_NOT_FOUND)
+      .render("404", { pageTitle: "Room not found." });
+  }
+  if (String(room.host) !== String(_id)) {
+    return res.status(HTTP_FORBIDDEN).redirect("/");
+  }
   // findByIdAndDelete(id) (=findOneAndDelete({_id:id}))
   await Room.findByIdAndDelete(id);
+  user.rooms.splice(user.rooms.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
